Add files that were on pip
Pip and this GitHub were not synchronized. In particular the code on pip worked for arbitrary data input because it read the dimension while the GitHub code assumed a given structure. I also fixed the warning of edge_color which in more mordern versions needs to be called border_color